refactor: creating IframeSidebar.js component to display chat view correctly with sidebar
